Changes to: Stores, other, File: Append_menus.py, other,
- Fixed bug allowing player to move while in combat,
- Added weapon testing,
- Removed <l> binding to battle,
- Added store function (see G tile in Room_2),
- code cleanup.

Note 1: "other" refers to all files in which only minor changes were made for the code cleanup.

Note 2: The tool for creating rooms mentioned a couple pushes ago is now ready for use. I might create a second repo for it once I deem it publish worthy, but as of now it still only has half the functions I want it to have.

Next is creating better visuals and preparing for demo release.