Changes to: Almost all. File: Almost all.
- Concluded code cleanup and fixed bugs,
- Repainted old buttons (and removed 'Test' from their names),
- Added some images for enemies,
- Added info card on click for enemies in battle,
- Added effects for barrier and all 3 passive spells,
- Slowed down combat and redesigned combat HUD,
- Added Control settings,
- Added Inventory/Stats menu (bound to I),
- Item stores are now saved in files, fixing the bug where world objects  were not reset upon loading files but on opening the game as stores are the only changeable world objects.
- Changed Demo_layout.txt to Guide.md, It's now a full explanation of the Demo.
- Attempted to add new player images but gave up halfway.

Note: I forgot to mention some changes in the previous push:
      - Added 4 unique enemies for the field,
      - Mild re-balancing off spells,
      - Added combat actions for defend and run.

Note 2: This was meant to be a small push, but I couldn't stop adding new things... If I have the time I might even add the other player images...

Note 3: To add the North-facing player images, add a 2 to the end of the path for player_idle_N.gif and player_movement_N.gif in Player_class.py. and change the zoom values to 5.

This is the last of two concluding pushes for the Demo edition 2022. After this, work on Destined will be done mostly offline.